DRY'd out Footer and built new footerTop and footerBottom components to clean up Footer file
